+++ src/task/task.controller.ts
@@ -1,5 +1,4 @@
 import { JwtAuthGuard } from '@/guard/jwt-auth.guard'
-import SerializerInterceptor from '@/interceptors/Serializer.interceptor'
 import { StageService } from '@/stage/stage.service'
 import {
 	BadRequestException,
@@ -13,11 +12,9 @@ import {
 	Patch,
 	Post,
 	UseGuards,
-	UseInterceptors,
 } from '@nestjs/common'
 import { CreateTaskDto } from './dto/create-task.dto'
 import { UpdateTaskDto } from './dto/update-task.dto'
-import { Task } from './entities/task.entity'
 import { TaskService } from './task.service'
 
 @Controller('task')
@@ -29,7 +26,6 @@ export class TaskController {
 	) {}
 
 	@UseGuards(JwtAuthGuard)
-	@UseInterceptors(SerializerInterceptor(Task))
 	@Post()
 	async create(@Body() createTaskDto: CreateTaskDto) {
 		try {
@@ -56,7 +52,6 @@ export class TaskController {
 	}
 
 	@UseGuards(JwtAuthGuard)
-	@UseInterceptors(SerializerInterceptor(Task))
 	@Patch(':id')
 	async update(@Param('id') id: string, @Body() updateTaskDto: UpdateTaskDto) {
 		try {
@@ -79,11 +74,10 @@ export class TaskController {
 	}
 
 	@UseGuards(JwtAuthGuard)
-	@UseInterceptors(SerializerInterceptor(Task))
 	@Post('/duplicate')
-	async duplicate(@Body() duplicateTaskDto: { id: string; stageId: string }) {
+	async duplicate(@Body() duplicateTaskDto: { id: string }) {
 		try {
-			const { id, stageId } = duplicateTaskDto
+			const { id } = duplicateTaskDto
 
 			const foundTask = await this.taskService.findById(id)
 
@@ -91,9 +85,21 @@ export class TaskController {
 				throw new BadRequestException('Task not found')
 			}
 
+			const foundStage = await this.stageService.findByTaskId(id)
+
+			if (!foundStage) {
+				throw new BadRequestException('Stage not found')
+			}
+
 			const createdTask = await this.taskService.duplicate(id)
 
-			await this.stageService.addTasks(stageId, [createdTask._id])
+			const foundIdx = foundStage.tasks.findIndex(
+				pred => pred.toString() === foundTask._id.toString(),
+			)
+
+			foundStage.tasks.splice(foundIdx + 1, 0, createdTask._id)
+
+			await foundStage.save()
 
 			return createdTask
 		} catch (e) {
@@ -105,7 +111,19 @@ export class TaskController {
 	@Delete(':id')
 	async remove(@Param('id') id: string) {
 		try {
+			const foundStage = await this.stageService.findByTaskId(id)
+
+			if (!foundStage) {
+				throw new BadRequestException('Stage not found')
+			}
+
 			const deleteResult = await this.taskService.remove(id)
+
+			const foudnIdx = foundStage.tasks.findIndex(
+				pred => pred.toString() === id,
+			)
+
+			foundStage.tasks.splice(foudnIdx, 1)
 
 			return {
 				success: !!deleteResult.deletedCount,
